build: Fix universal rendering of component library
