+++ .circleci/config.yml
@@ -65,7 +65,7 @@ jobs:
           <<: *set_env
       - restore_cache:
           <<: *npm_cache
-      - run: yarn format:check ${AFFECTED_ARGS}
+      - run: yarn format:check --base=master
   lint:
     <<: *defaults
     steps:
@@ -77,7 +77,7 @@ jobs:
       - restore_cache:
           <<: *npm_cache
       - run: ./node_modules/.bin/nx workspace-lint
-      - run: yarn affected:lint ${AFFECTED_ARGS} --parallel
+      - run: yarn affected:lint --base=master --parallel
   build:
     <<: *defaults
     steps:
@@ -88,7 +88,7 @@ jobs:
           <<: *set_env
       - restore_cache:
           <<: *npm_cache
-      - run: yarn affected:build -- ${AFFECTED_ARGS} --parallel -- --configuration production
+      - run: yarn affected:build --base=master --parallel --configuration=production
       - save_cache:
           key: build-{{ .Environment.CIRCLE_WORKFLOW_ID }}
           paths:
@@ -105,7 +105,7 @@ jobs:
           <<: *set_env
       - restore_cache:
           <<: *npm_cache
-      - run: yarn affected:test -- ${AFFECTED_ARGS} --parallel -- --ci --code-coverage
+      - run: yarn affected:test --base=master --parallel
 
 
 workflows:
