+++ apps/universal/src/app/app.server.module.ts
@@ -1,0 +1,21 @@
+import {
+  ServerModule,
+  ServerTransferStateModule,
+} from '@angular/platform-server';
+
+import { AppModule } from './app.module';
+import { KitchenSink } from './kitchen-sink/kitchen-sink';
+// import { DtIconModule } from '@dynatrace/barista-components/icon';
+import { ModuleMapLoaderModule } from '@nguniversal/module-map-ngfactory-loader';
+import { NgModule } from '@angular/core';
+
+@NgModule({
+  imports: [
+    AppModule,
+    ServerModule,
+    ServerTransferStateModule,
+    ModuleMapLoaderModule,
+  ],
+  bootstrap: [KitchenSink],
+})
+export class AppServerModule {}
