+++ apps/universal/src/app/kitchen-sink/kitchen-sink.html
@@ -253,7 +253,7 @@
 <dt-empty-state>
   <dt-empty-state-item>
     <dt-empty-state-item-img>
-      <img src="/assets/cta-noagent.svg" alt="No agent" />
+      <img src="/assets/icons/agent.svg" alt="No agent" />
     </dt-empty-state-item-img>
 
     <dt-empty-state-item-title>
@@ -267,7 +267,7 @@
 
   <dt-empty-state-item>
     <dt-empty-state-item-img>
-      <img src="/assets/cta-noagent.svg" alt="No agent" />
+      <img src="/assets/icons/agent.svg" alt="No agent" />
     </dt-empty-state-item-img>
 
     <dt-empty-state-item-title>
@@ -281,7 +281,7 @@
 
   <dt-empty-state-item>
     <dt-empty-state-item-img>
-      <img src="/assets/cta-noagent.svg" alt="No agent" />
+      <img src="/assets/icons/agent.svg" alt="No agent" />
     </dt-empty-state-item-img>
 
     <dt-empty-state-item-title>
