+++ apps/universal/src/app/kitchen-sink/kitchen-sink.ts
@@ -1,0 +1,168 @@
+import {
+  DtTreeControl,
+  DtTreeDataSource,
+  DtTreeFlattener,
+} from '@dynatrace/barista-components/core';
+
+import { Component } from '@angular/core';
+import { DtIconType } from '@dynatrace/dt-iconpack';
+
+const TESTDATA: ThreadNode[] = [
+  {
+    name: 'hz.hzInstance_1_cluster.thread',
+    icon: 'apache-tomcat',
+    threadlevel: 'S0',
+    totalTimeConsumption: 150,
+    waiting: 123,
+    running: 20,
+    blocked: 0,
+    children: [
+      {
+        name:
+          'hz.hzInstance_1_cluster.thread_1_hz.hzInstance_1_cluster.thread-1',
+        icon: 'apache-tomcat',
+        threadlevel: 'S1',
+        totalTimeConsumption: 150,
+        waiting: 123,
+        running: 20,
+        blocked: 0,
+      },
+      {
+        name: 'hz.hzInstance_1_cluster.thread-2',
+        icon: 'apache-tomcat',
+        threadlevel: 'S1',
+        totalTimeConsumption: 150,
+        waiting: 130,
+        running: 0,
+        blocked: 0,
+      },
+    ],
+  },
+  {
+    name: 'jetty',
+    icon: 'apache-tomcat',
+    threadlevel: 'S0',
+    totalTimeConsumption: 150,
+    waiting: 123,
+    running: 20,
+    blocked: 0,
+    children: [
+      {
+        name: 'jetty-422',
+        icon: 'apache-tomcat',
+        threadlevel: 'S1',
+        totalTimeConsumption: 150,
+        waiting: 123,
+        running: 20,
+        blocked: 0,
+      },
+      {
+        name: 'jetty-423',
+        icon: 'apache-tomcat',
+        threadlevel: 'S1',
+        totalTimeConsumption: 150,
+        waiting: 130,
+        running: 0,
+        blocked: 0,
+      },
+      {
+        name: 'jetty-424',
+        icon: 'apache-tomcat',
+        threadlevel: 'S1',
+        totalTimeConsumption: 150,
+        waiting: 130,
+        running: 0,
+        blocked: 0,
+      },
+    ],
+  },
+  {
+    name: 'Downtime timer',
+    icon: 'apache-tomcat',
+    threadlevel: 'S0',
+    totalTimeConsumption: 150,
+    waiting: 123,
+    running: 20,
+    blocked: 0,
+  },
+];
+
+export class ThreadNode {
+  name: string;
+  threadlevel: string;
+  totalTimeConsumption: number;
+  blocked: number;
+  running: number;
+  waiting: number;
+  icon: DtIconType;
+  children?: ThreadNode[];
+}
+
+export class ThreadFlatNode {
+  name: string;
+  threadlevel: string;
+  totalTimeConsumption: number;
+  blocked: number;
+  running: number;
+  waiting: number;
+  icon: DtIconType;
+  level: number;
+  expandable: boolean;
+}
+
+@Component({
+  selector: 'dt-kitchen-sink',
+  templateUrl: './kitchen-sink.html',
+})
+export class KitchenSink {
+  tableDataSource: object[] = [
+    { host: 'et-demo-2-win4' },
+    { host: 'et-demo-2-win6' },
+    { host: 'et-demo-2-win8' },
+  ];
+
+  treeControl: DtTreeControl<ThreadFlatNode>;
+  treeFlattener: DtTreeFlattener<ThreadNode, ThreadFlatNode>;
+  treeTableDataSource: DtTreeDataSource<ThreadNode, ThreadFlatNode>;
+
+  constructor() {
+    this.treeControl = new DtTreeControl<ThreadFlatNode>(
+      this._getLevel,
+      this._isExpandable,
+    );
+    this.treeFlattener = new DtTreeFlattener(
+      this.transformer,
+      this._getLevel,
+      this._isExpandable,
+      this._getChildren,
+    );
+    this.treeTableDataSource = new DtTreeDataSource(
+      this.treeControl,
+      this.treeFlattener,
+    );
+    this.treeTableDataSource.data = TESTDATA;
+  }
+
+  hasChild = (_: number, _nodeData: ThreadFlatNode) => _nodeData.expandable;
+
+  transformer = (node: ThreadNode, level: number): ThreadFlatNode => {
+    const flatNode = new ThreadFlatNode();
+    flatNode.name = node.name;
+    flatNode.level = level;
+    flatNode.threadlevel = node.threadlevel;
+    flatNode.expandable = !!node.children;
+    flatNode.blocked = node.blocked;
+    flatNode.running = node.running;
+    flatNode.waiting = node.waiting;
+    flatNode.totalTimeConsumption = node.totalTimeConsumption;
+    flatNode.icon = node.icon;
+    return flatNode;
+  };
+
+  private _getLevel = (node: ThreadFlatNode) => node.level;
+
+  private _isExpandable = (node: ThreadFlatNode) => node.expandable;
+
+  private _getChildren = (node: ThreadNode): ThreadNode[] =>
+    node.children || [];
+}
