+++ apps/universal/src/index.html
@@ -2,8 +2,8 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <title>Dynatrace Angular Components Univeral Kitchen Sink Test</title>
     <base href="/" />
+    <title>Barista Components Universal</title>
   </head>
   <body>
     <dt-kitchen-sink>Loading...</dt-kitchen-sink>
