--- apps/universal/src/kitchen-sink/kitchen-sink.ts
@@ -1,259 +1,0 @@
-import { Component, NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { ServerModule } from '@angular/platform-server';
-import { DtIconType } from '@dynatrace/dt-iconpack';
-
-import { DtAlertModule } from '@dynatrace/barista-components/alert';
-import { DtAutocompleteModule } from '@dynatrace/barista-components/autocomplete';
-import { DtButtonModule } from '@dynatrace/barista-components/button';
-import { DtButtonGroupModule } from '@dynatrace/barista-components/button-group';
-import { DtCardModule } from '@dynatrace/barista-components/card';
-import { DtCheckboxModule } from '@dynatrace/barista-components/checkbox';
-import { DtConsumptionModule } from '@dynatrace/barista-components/consumption';
-import { DtContainerBreakpointObserverModule } from '@dynatrace/barista-components/container-breakpoint-observer';
-import { DtContextDialogModule } from '@dynatrace/barista-components/context-dialog';
-import { DtCopyToClipboardModule } from '@dynatrace/barista-components/copy-to-clipboard';
-import {
-  DtTreeControl,
-  DtTreeDataSource,
-  DtTreeFlattener,
-} from '@dynatrace/barista-components/core';
-import { DtDrawerModule } from '@dynatrace/barista-components/drawer';
-import { DtEmptyStateModule } from '@dynatrace/barista-components/empty-state';
-import { DtEventChartModule } from '@dynatrace/barista-components/event-chart';
-import { DtExpandableTextModule } from '@dynatrace/barista-components/expandable-text';
-import { DtFilterFieldModule } from '@dynatrace/barista-components/filter-field';
-import { DtHighlightModule } from '@dynatrace/barista-components/highlight';
-import { DtIconModule } from '@dynatrace/barista-components/icon';
-import { DtInfoGroupModule } from '@dynatrace/barista-components/info-group';
-import { DtInputModule } from '@dynatrace/barista-components/input';
-import { DtLoadingDistractorModule } from '@dynatrace/barista-components/loading-distractor';
-import { DtMenuModule } from '@dynatrace/barista-components/menu';
-import { DtOverlayModule } from '@dynatrace/barista-components/overlay';
-import { DtPaginationModule } from '@dynatrace/barista-components/pagination';
-import { DtProgressBarModule } from '@dynatrace/barista-components/progress-bar';
-import { DtProgressCircleModule } from '@dynatrace/barista-components/progress-circle';
-import { DtRadioModule } from '@dynatrace/barista-components/radio';
-import { DtSelectModule } from '@dynatrace/barista-components/select';
-import { DtShowMoreModule } from '@dynatrace/barista-components/show-more';
-import { DtStepperModule } from '@dynatrace/barista-components/stepper';
-import { DtSwitchModule } from '@dynatrace/barista-components/switch';
-import { DtTableModule } from '@dynatrace/barista-components/table';
-import { DtTagModule } from '@dynatrace/barista-components/tag';
-import { DtTileModule } from '@dynatrace/barista-components/tile';
-import { DtTimelineChartModule } from '@dynatrace/barista-components/timeline-chart';
-import { DtToggleButtonGroupModule } from '@dynatrace/barista-components/toggle-button-group';
-import { DtTopBarNavigationModule } from '@dynatrace/barista-components/top-bar-navigation';
-import { DtTreeTableModule } from '@dynatrace/barista-components/tree-table';
-
-const TESTDATA: ThreadNode[] = [
-  {
-    name: 'hz.hzInstance_1_cluster.thread',
-    icon: 'apache-tomcat',
-    threadlevel: 'S0',
-    totalTimeConsumption: 150,
-    waiting: 123,
-    running: 20,
-    blocked: 0,
-    children: [
-      {
-        name:
-          'hz.hzInstance_1_cluster.thread_1_hz.hzInstance_1_cluster.thread-1',
-        icon: 'apache-tomcat',
-        threadlevel: 'S1',
-        totalTimeConsumption: 150,
-        waiting: 123,
-        running: 20,
-        blocked: 0,
-      },
-      {
-        name: 'hz.hzInstance_1_cluster.thread-2',
-        icon: 'apache-tomcat',
-        threadlevel: 'S1',
-        totalTimeConsumption: 150,
-        waiting: 130,
-        running: 0,
-        blocked: 0,
-      },
-    ],
-  },
-  {
-    name: 'jetty',
-    icon: 'apache-tomcat',
-    threadlevel: 'S0',
-    totalTimeConsumption: 150,
-    waiting: 123,
-    running: 20,
-    blocked: 0,
-    children: [
-      {
-        name: 'jetty-422',
-        icon: 'apache-tomcat',
-        threadlevel: 'S1',
-        totalTimeConsumption: 150,
-        waiting: 123,
-        running: 20,
-        blocked: 0,
-      },
-      {
-        name: 'jetty-423',
-        icon: 'apache-tomcat',
-        threadlevel: 'S1',
-        totalTimeConsumption: 150,
-        waiting: 130,
-        running: 0,
-        blocked: 0,
-      },
-      {
-        name: 'jetty-424',
-        icon: 'apache-tomcat',
-        threadlevel: 'S1',
-        totalTimeConsumption: 150,
-        waiting: 130,
-        running: 0,
-        blocked: 0,
-      },
-    ],
-  },
-  {
-    name: 'Downtime timer',
-    icon: 'apache-tomcat',
-    threadlevel: 'S0',
-    totalTimeConsumption: 150,
-    waiting: 123,
-    running: 20,
-    blocked: 0,
-  },
-];
-
-export class ThreadNode {
-  name: string;
-  threadlevel: string;
-  totalTimeConsumption: number;
-  blocked: number;
-  running: number;
-  waiting: number;
-  icon: DtIconType;
-  children?: ThreadNode[];
-}
-
-export class ThreadFlatNode {
-  name: string;
-  threadlevel: string;
-  totalTimeConsumption: number;
-  blocked: number;
-  running: number;
-  waiting: number;
-  icon: DtIconType;
-  level: number;
-  expandable: boolean;
-}
-
-@Component({
-  selector: 'dt-kitchen-sink',
-  templateUrl: './kitchen-sink.html',
-})
-export class KitchenSink {
-  tableDataSource: object[] = [
-    { host: 'et-demo-2-win4' },
-    { host: 'et-demo-2-win6' },
-    { host: 'et-demo-2-win8' },
-  ];
-
-  treeControl: DtTreeControl<ThreadFlatNode>;
-  treeFlattener: DtTreeFlattener<ThreadNode, ThreadFlatNode>;
-  treeTableDataSource: DtTreeDataSource<ThreadNode, ThreadFlatNode>;
-
-  constructor() {
-    this.treeControl = new DtTreeControl<ThreadFlatNode>(
-      this._getLevel,
-      this._isExpandable,
-    );
-    this.treeFlattener = new DtTreeFlattener(
-      this.transformer,
-      this._getLevel,
-      this._isExpandable,
-      this._getChildren,
-    );
-    this.treeTableDataSource = new DtTreeDataSource(
-      this.treeControl,
-      this.treeFlattener,
-    );
-    this.treeTableDataSource.data = TESTDATA;
-  }
-
-  hasChild = (_: number, _nodeData: ThreadFlatNode) => _nodeData.expandable;
-
-  transformer = (node: ThreadNode, level: number): ThreadFlatNode => {
-    const flatNode = new ThreadFlatNode();
-    flatNode.name = node.name;
-    flatNode.level = level;
-    flatNode.threadlevel = node.threadlevel;
-    flatNode.expandable = !!node.children;
-    flatNode.blocked = node.blocked;
-    flatNode.running = node.running;
-    flatNode.waiting = node.waiting;
-    flatNode.totalTimeConsumption = node.totalTimeConsumption;
-    flatNode.icon = node.icon;
-    return flatNode;
-  };
-
-  private _getLevel = (node: ThreadFlatNode) => node.level;
-
-  private _isExpandable = (node: ThreadFlatNode) => node.expandable;
-
-  private _getChildren = (node: ThreadNode): ThreadNode[] =>
-    node.children || [];
-}
-
-@NgModule({
-  imports: [
-    BrowserModule.withServerTransition({ appId: 'kitchen-sink' }),
-    DtAlertModule,
-    DtAutocompleteModule,
-    DtButtonModule,
-    DtCheckboxModule,
-    DtTableModule,
-    DtLoadingDistractorModule,
-    DtTileModule,
-    DtTagModule,
-    DtCardModule,
-    DtContextDialogModule,
-    DtDrawerModule,
-    DtCopyToClipboardModule,
-    DtButtonGroupModule,
-    DtIconModule.forRoot({ svgIconLocation: '/lib/assets/icons/{{name}}.svg' }),
-    DtRadioModule,
-    DtShowMoreModule,
-    DtProgressCircleModule,
-    DtPaginationModule,
-    DtSwitchModule,
-    DtProgressBarModule,
-    DtSelectModule,
-    DtInputModule,
-    DtOverlayModule,
-    DtTreeTableModule,
-    DtToggleButtonGroupModule,
-    DtInfoGroupModule,
-    DtHighlightModule,
-    DtConsumptionModule,
-    DtFilterFieldModule,
-    DtMenuModule,
-    DtEmptyStateModule,
-    DtTimelineChartModule,
-    DtExpandableTextModule,
-    DtEventChartModule,
-    DtTopBarNavigationModule,
-    DtStepperModule,
-    DtContainerBreakpointObserverModule,
-  ],
-  bootstrap: [KitchenSink],
-  declarations: [KitchenSink],
-})
-export class KitchenSinkClientModule {}
-
-@NgModule({
-  imports: [KitchenSinkClientModule, ServerModule],
-  bootstrap: [KitchenSink],
-})
-export class KitchenSinkServerModule {}
