+++ apps/universal/src/main.server.ts
@@ -1,0 +1,5 @@
+import { enableProdMode } from '@angular/core';
+
+enableProdMode();
+
+export { AppServerModule } from './app/app.server.module';
