+++ apps/universal/src/main.ts
@@ -1,1 +1,13 @@
-export * from './kitchen-sink/kitchen-sink';
+import { AppModule } from './app/app.module';
+import { enableProdMode } from '@angular/core';
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+
+enableProdMode();
+
+document.addEventListener('DOMContentLoaded', () => {
+  platformBrowserDynamic()
+    .bootstrapModule(AppModule)
+    .catch(err => {
+      console.error(err);
+    });
+});
