--- apps/universal/tsconfig-prerender.json
@@ -1,24 +1,0 @@
-// TypeScript config file that is used to compile the files that prerender the Universal app.
-{
-  "compilerOptions": {
-    "declaration": false,
-    "stripInternal": false,
-    "experimentalDecorators": true,
-    "noUnusedParameters": true,
-    "strictNullChecks": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "outDir": ".",
-    "rootDir": ".",
-    "sourceMap": true,
-    "target": "es2015",
-    "lib": ["es2015", "dom"],
-    "skipLibCheck": true,
-    "types": [],
-    "baseUrl": ".",
-    "paths": {
-      "@dynatrace/barista-components/*": ["./lib/*"]
-    }
-  },
-  "files": ["prerender.ts"]
-}
