+++ apps/universal/tsconfig.app.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "outDir": "../../dist/out-tsc",
+    "types": []
+  },
+  "exclude": ["src/server.ts", "**/*.spec.ts"],
+  "include": ["**/*.ts"]
+}
