+++ package.json
@@ -24,7 +24,6 @@
     "stylelint-check": "stylelint-config-prettier-check",
     "prettier": "prettier \"./**/*.{ts,js,css,scss,json,html,md}\"",
     "prettier:fix": "npm run prettier --write",
-    "universal": "gulp universal",
     "schematics:build": "gulp schematics:build",
     "bitbucket-tools": "nodemon --config ./tools/bitbucket-tools/nodemon.json",
     "bitbucket-tools:test": "jest --rootDir tools/bitbucket-tools/",
@@ -85,6 +84,8 @@
     "@angular/core": "^8.0.0",
     "@angular/forms": "^8.0.0",
     "@angular/platform-browser": "^8.0.0",
+    "@nguniversal/express-engine": "^8.1.1",
+    "@nguniversal/module-map-ngfactory-loader": "^8.1.1",
     "core-js": "^3.1.3",
     "d3-scale": "^3.0.0",
     "d3-shape": "^1.3.5",
